refactor(rust/core): hoist wire-agnostic session/channel logic into core
Move the channel store, voucher signature verification, and the
voucher-accept state machine out of `solana-mpp` and into
`solana-pay-core`, so both mpp and x402 build on one tested
implementation (mirrors the earlier `payment_channels` / `parse_units`
extractions). `solana-x402` deliberately does not depend on `solana-mpp`,
so this is the prerequisite for the x402 `batch-settlement` scheme.

- core::store        <- mpp/src/store.rs (ChannelStore/ChannelState/
                        MemoryChannelStore, ...); mpp::store re-exports it
- core::voucher      <- verify_voucher_signature (Ed25519, ed25519-dalek)
- core::session      <- accept_voucher (monotonic + deposit cap +
                        idempotent replay + atomic compare-and-set advance)

mpp::server::session now delegates to these helpers, keeping the MPP-wire
mapping and metering in mpp. Guardrail: mpp's full suite stays green.
